구독

+++ subscription-service/src/main/java/com/bit/docker/subscriptionservice/controller/InternalSubscriptionController.java
@@ -16,20 +16,20 @@ public class InternalSubscriptionController {
 
     // fanout용: 개인(아이돌) 구독자 userId 리스트
     @GetMapping("/idols/{idolId}/user-ids")
-    public ResponseEntity<List<Integer>> getIdolSubscriberUserIds(@PathVariable("idolId") Long idolId) {
+    public ResponseEntity<List<Integer>> getIdolSubscriberUserIds(@PathVariable("idolId") int idolId) {
         return ResponseEntity.ok(subscriptionService.getActiveSubscriberUserIdsByIdolId(idolId));
     }
 
     // fanout용: 그룹 구독자 userId 리스트
     @GetMapping("/groups/{groupId}/user-ids")
-    public ResponseEntity<List<Integer>> getGroupSubscriberUserIds(@PathVariable("groupId") Long groupId) {
+    public ResponseEntity<List<Integer>> getGroupSubscriberUserIds(@PathVariable("groupId") int groupId) {
         return ResponseEntity.ok(subscriptionService.getActiveSubscriberUserIdsByGroupId(groupId));
     }
 
     // idol 게시판 상세 열람용: userId가 idolId 유료구독(active) 중인지
     @GetMapping("/idols/{idolId}/users/{userId}/active")
     public ResponseEntity<Boolean> isActiveIdolSubscriber(
-            @PathVariable("idolId") Long idolId,
+            @PathVariable("idolId") int idolId,
             @PathVariable("userId") int userId
     ) {
         return ResponseEntity.ok(subscriptionService.isActiveIdolSubscriber(userId, idolId));
